GaussMLE

Documentation for GaussMLE.

GaussMLE.center_of_mass_2d

GaussMLE.compute_alpha

GaussMLE.derivative_alpha_z

GaussMLE.derivative_integral_gaussian_1d

GaussMLE.derivative_integral_gaussian_1d_sigma

GaussMLE.derivative_integral_gaussian_2d_sigma

GaussMLE.derivative_integral_gaussian_2d_z

GaussMLE.gaussian_max_min_2d

GaussMLE.integral_gaussian_1d

GaussMLE.matrix_inverse!

GaussMLE.second_derivative_alpha_z

GaussMLE.center_of_mass_2d — Method

center_of_mass_2d(sz::Int, data::Array{<:Real})

Compute the center of mass of a square 2D data array.

GaussMLE.compute_alpha — Method

compute_alpha(z::Real, Ax::Real, Bx::Real, d::Real)

Compute the alpha value based on the given parameters.

GaussMLE.derivative_alpha_z — Method

derivative_alpha_z(z::Real, Ax::Real, Bx::Real, d::Real)

Compute the derivative of alpha with respect to z.

GaussMLE.derivative_integral_gaussian_1d — Method

derivative_integral_gaussian_1d(ii::Int, x::Real, sigma::Real, N::Real, PSFy::Real)

Compute the derivative of the integral of a 1D Gaussian function.

GaussMLE.derivative_integral_gaussian_1d_sigma — Method

derivative_integral_gaussian_1d_sigma(ii::Int, x::Real, Sx::Real, N::Real, PSFy::Real)

Compute the derivative of the integral of a 1D Gaussian function with respect to sigma.

GaussMLE.derivative_integral_gaussian_2d_sigma — Method

derivative_integral_gaussian_2d_sigma(ii::Int, jj::Int, x::Real, y::Real, S::Real, N::Real, PSFx::Real, PSFy::Real)

Compute the derivative of the integral of a 2D Gaussian function with respect to sigma.

GaussMLE.derivative_integral_gaussian_2d_z — Method

derivative_integral_gaussian_2d_z(ii::Int, jj::Int, theta, PSFSigma_x::Real, PSFSigma_y::Real, Ax::Real, Ay::Real, Bx::Real, By::Real, gamma::Real, d::Real, dudt, d2udt2)

Compute the derivative of the integral of a 2D Gaussian function with respect to z.

GaussMLE.gaussian_max_min_2d — Method

gaussian_max_min_2d(sz::Int, sigma::Real, data::Array{<:Real})

Compute the maximum and minimum values after applying a Gaussian filter to a 2D data array.

GaussMLE.integral_gaussian_1d — Method

integral_gaussian_1d(ii::Int, position::Real, sigma::Real)

Calculate the integral of a 1D Gaussian function.

GaussMLE.matrix_inverse! — Method

matrix_inverse!(m::Array{T, 2}, sz::Int) where T <: Real

Compute the inverse of a matrix m with size sz.

This implementation is designed to be compatible with both CPU and GPU computations. Note: This is a custom implementation. Consider using Julia's built-in inv function for standard use-cases on the CPU.

This function modifies the input matrix m in place as part of the computation. If the original matrix data needs to be preserved, make sure to provide a copy of the matrix to this function.

Arguments

m: Input matrix to be inverted. This matrix will be modified in place.

sz: The size of the matrix.

Returns

m_inverse: The inverse of the input matrix.

m_inverse_diag: The diagonal elements of the inverse matrix.

GaussMLE.second_derivative_alpha_z — Method

second_derivative_alpha_z(z::Real, Ax::Real, Bx::Real, d::Real)

Compute the second derivative of alpha with respect to z.