GaussMLE.GaussLib

Overview

GaussMLE.GaussLib — Module

GaussLib

Low level functions for calcualting the Gaussian model and its derivatives.

API

GaussMLE.GaussLib.center_of_mass_2d

GaussMLE.GaussLib.compute_alpha

GaussMLE.GaussLib.derivative_alpha_z

GaussMLE.GaussLib.derivative_integral_gaussian_1d

GaussMLE.GaussLib.derivative_integral_gaussian_1d_sigma

GaussMLE.GaussLib.derivative_integral_gaussian_2d_sigma

GaussMLE.GaussLib.gaussian_max_min_2d

GaussMLE.GaussLib.integral_gaussian_1d

GaussMLE.GaussLib.second_derivative_alpha_z

GaussMLE.GaussLib.center_of_mass_2d — Method

center_of_mass_2d(sz::Int, data::Array{T}) where T <: Real

Compute the center of mass of a square 2D data array.

GaussMLE.GaussLib.compute_alpha — Method

compute_alpha(z::T, Ax::T, Bx::T, d::T) where T <: Real

Compute the alpha value based on the given parameters.

GaussMLE.GaussLib.derivative_alpha_z — Method

derivative_alpha_z(z::T, Ax::T, Bx::T, d::T) where T <: Real

Compute the derivative of alpha with respect to z.

GaussMLE.GaussLib.derivative_integral_gaussian_1d — Method

derivative_integral_gaussian_1d(ii::Int, x::T, sigma::T, N::T, PSFy::T) where T <: Real

Compute the derivative of the integral of a 1D Gaussian function with respect to x.

GaussMLE.GaussLib.derivative_integral_gaussian_1d_sigma — Method

derivative_integral_gaussian_1d_sigma(i::Int, x::T, Sx::T, N::T, PSFy::T) where T <: Real

Compute the derivative of the integral of a 1D Gaussian function with respect to sigma.

GaussMLE.GaussLib.derivative_integral_gaussian_2d_sigma — Method

derivative_integral_gaussian_2d_sigma(i::Int, j::Int, x::T, y::T, S::T, N::T, PSFx::T, PSFy::T) where T <: Real

Compute the derivative of the integral of a 2D Gaussian function with respect to sigma.

GaussMLE.GaussLib.gaussian_max_min_2d — Method

gaussian_max_min_2d(sz::Int, sigma::T, data::Array{T}) where T <: Real

Compute the maximum and minimum values after applying a Gaussian filter to a 2D data array.

GaussMLE.GaussLib.integral_gaussian_1d — Method

integral_gaussian_1d(ii::Int, position::T, sigma::T) where T <: Real

Calculate the integral of a 1D Gaussian function.

GaussMLE.GaussLib.second_derivative_alpha_z — Method

second_derivative_alpha_z(z::T, Ax::T, Bx::T, d::T) where T <: Real

Compute the second derivative of alpha with respect to z.