GaussMLE.GaussSim

Overview

GaussMLE.GaussSim — Module

GaussSim

Methods for creating simulated data for testing and benchmarking.

API

GaussMLE.GaussSim.genroi!

GaussMLE.GaussSim.genstack

GaussMLE.GaussSim.genstack

GaussMLE.GaussSim.poissonnoise!

GaussMLE.GaussSim.genroi! — Method

poissonnoise!(stack)

Add Poisson noise to a 3D stack of images.

stack::Array{T}: The 3D stack to which Poisson noise will be added.

The function modifies stack in-place and returns nothing.

GaussMLE.GaussSim.genstack — Method

genstack(boxsize, stackdepth, modelsymbol; T=Float32, poissonnoise=false)

Generate a 3D stack of images based on a pre-defined Gaussian model identified by modelsymbol.

Default values will be used for the model paramters.

Arguments

boxsize::Int: The size of each side of the square 2D ROI.

stackdepth::Int: The number of slices in the 3D stack.

modelsymbol::Symbol: A symbol identifying the Gaussian model to use.

T::Type{<:Real} (optional): The type of real numbers to use (default is Float32).

poissonnoise::Bool (optional): Whether to add Poisson noise to the stack (default is false).

Returns

A tuple containing the generated 3D stack, the vector of model parameters for each slice, and the model-specific arguments.

GaussMLE.GaussSim.genstack — Method

genstack(boxsize, stackdepth, θ, args; poissonnoise=false)

Generate a 3D stack of images based on the given Gaussian model parameters θ and additional arguments args.

Each slice in the stack is a 2D ROI generated by the model.

Arguments

boxsize::Int: The size of each side of the square 2D ROI.

stackdepth::Int: The number of slices in the 3D stack.

θ::Vector{<:GaussMLEParams{T}}: A vector of instances of types that inherit from GaussMLEParams containing the model parameters for each slice.

args::GaussMLEArgs{T}: An instance of a type that inherits from GaussMLEArgs containing model-specific arguments.

poissonnoise::Bool (optional): Whether to add Poisson noise to the stack (default is false).

Returns

The generated 3D stack.

GaussMLE.GaussSim.poissonnoise! — Method

poissonnoise!(stack)

Add Poisson noise to a 3D stack of images.

Arguments

stack::Array{T}: The 3D stack to which Poisson noise will be added.

Returns

Modifies stack in-place and returns nothing.